fix: extract symlink_test.go from all_test.go and separate it on os in symlink_test_x.go
